VTMD: configure prelocking for ..as of.. stmt
Problem:
When nested queries are used inside `as of` statement, needed tables
are locked before execution, but whole thread tables are unlocked after
it. So, we unlock tables that shouldn't be unlocked. This affects `select`
and `show create table` queries.

Solution:
Nested locking is forbidden. Instead, all tables are to be opened at
once. So we add all vtmd tables to lexer's table list and then open
them in a row with other tables used in query.